fix: refactor main
